% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles-package.R
\name{measles_cpp_version}
\alias{measles_cpp_version}
\title{Version of the measles C++ code}
\usage{
measles_cpp_version()
}
\value{
A character string representing the version of the C++ library.
}
\description{
Returns the version of the C++ library measles. The code
is hosted on GitHub at \url{https://github.com/UofUEpiBio/epiworld},
and it is part of the epiworld C++ library.
}
\examples{
measles_cpp_version()

}
