% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{southwest_utah}
\alias{southwest_utah}
\title{Southwest Utah Population Data by Age Group}
\format{
A data frame with 18 rows and 4 columns:
\describe{
\item{age_labels}{character. Labels describing the age groups.}
\item{agepops}{numeric. Population counts for each age group.}
\item{agelims}{numeric. Age limit boundaries for each group.}
\item{vacc_rate}{numeric. Vaccination rate for each age group.}
}
}
\source{
The data was generated using the \code{multigroup.vaccine} R package:
Toth D (2025). \emph{multigroup.vaccine: Multigroup Vaccine Model}. R
package version 0.1.0, commit 54acee568fbb0666231ab59f19d33841337d7402,
\url{https://github.com/EpiForeSITE/multigroup-vaccine}. The source code
is available at \url{https://github.com/UofUEpiBio/measles}
}
\usage{
southwest_utah
}
\description{
A dataset containing population information for southwest Utah organized by age groups. \strong{IMPORTANT: THE SCHOOL-RELATED DATA IS FAKE. WE ARE STILL
DEVELOPING IT.}
}
\details{
This dataset provides demographic information for southwest Utah,
with population data disaggregated by 18 age categories. This
dataset matches the \link{southwest_utah_matrix} matrix.
}
\examples{
data(southwest_utah)
head(southwest_utah)

}
\keyword{datasets}
