% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_mixing_model.R
\name{calibrate_mixing_model}
\alias{calibrate_mixing_model}
\title{Calibrate a mixing model to a target reproduction number}
\usage{
calibrate_mixing_model(
  contact_matrix,
  target_rep_number,
  infectious_period_days,
  transmission_prob,
  ...
)
}
\arguments{
\item{contact_matrix}{A contact matrix to be calibrated.}

\item{target_rep_number}{The target reproduction number to calibrate to.}

\item{infectious_period_days}{The average number of days an individual is
infectious.}

\item{transmission_prob}{The probability of transmission per contact.}

\item{...}{Additional arguments to pass to \code{\link[stats:optimize]{stats::optimize()}}.}
}
\value{
The scaling factor for the contact matrix that achieves the target
reproduction number.
}
\description{
Uses the \code{\link[epiworldR:compute_reproduction_number]{epiworldR::compute_reproduction_number()}} function to
optimize the scaling factor of the contact matrix to match a target
reproduction number.
}
\examples{
data(short_creek_matrix, package = "measles")

# Calibrating for a measles model with R0 of 10
# assuming agents are infectious during prodomal stage
# (isolated during rash stage)
calibrate_mixing_model(
  contact_matrix = short_creek_matrix,
  target_rep_number = 10,
  infectious_period_days = 4,
  transmission_prob = 0.2
)

# You can then use the scaling factor in a mixing
# model. Instead of using the original contact matrix,
# you would use:
#   contact_matrix * scaling_factor
}
