% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterventionMeaslesPEP.R
\name{InterventionMeaslesPEP}
\alias{InterventionMeaslesPEP}
\title{Create a measles post-exposure prophylaxis (PEP) intervention}
\usage{
InterventionMeaslesPEP(
  name,
  mmr_efficacy,
  ig_efficacy,
  ig_half_life_mean,
  ig_half_life_sd,
  pep_willingness,
  mmr_window,
  target_states,
  states_if_pep_effective,
  states_if_pep_ineffective
)
}
\arguments{
\item{name}{Name of the intervention.}

\item{mmr_efficacy}{Probability of MMR vaccine efficacy.}

\item{ig_efficacy}{Probability of immunoglobulin (IG) efficacy.}

\item{ig_half_life_mean}{Mean of the half-life of immunoglobulin (IG) in days.}

\item{ig_half_life_sd}{Standard deviation of the half-life of immunoglobulin (IG) in days.}

\item{pep_willingness}{Probability that an individual will accept PEP.}

\item{mmr_window}{Time window for MMR vaccine administration.}

\item{target_states, states_if_pep_effective, states_if_pep_ineffective}{Integer vectors of target and destination states (see details).}
}
\value{
An object of class \code{epiworld_globalevent} representing the measles PEP
intervention.
}
\description{
Create a measles post-exposure prophylaxis (PEP) intervention
}
\details{
This functions creates a global event that represents a post-exposure
prophylaxis (PEP) intervention for measles. The intervention includes the
administration of MMR vaccine and immunoglobulin (IG) to individuals after
exposure to the virus, with the goal of reducing the probability
of infection and preventing the spread of the disease.

The process involves both PEP Measles-Mumps-Rubella (MMR) vaccine and
immunoglobulin (IG). The system decides which agent gets MMR or IG
based on the time since exposure and the willingness to accept PEP.
The flow is the following:
\enumerate{
\item Agents in \code{target_states} are eligible for PEP if they are
willing to accept it (based on \code{pep_willingness}).
\item If the agent is already infected (for example, in a latent state)
for at most \code{mmr_window} days, they are offered MMR vaccine.
Otherwise, they are offered IG.
\item Susceptible agents are offered the MMR vaccine, and if they accept,
they are automatically moved out of the quarantine process.
\item Agents who were already infected and got either MMR or IG may move
out of the quarantine process if the PEP is effective (based on
\code{mmr_efficacy} or \code{ig_efficacy}). The destination state depends
on whether the PEP was effective or not, and is determined by
\code{states_if_pep_effective} and \code{states_if_pep_ineffective},
respectively.
}

Since IG winds down over time, the IG "tool" may be removed from
the agent as a function of the half-life of IG (based on
\code{ig_half_life_mean} and \code{ig_half_life_sd}). Particularly, after
applied, the IG "tool" will have a random duration based on a normal
distribution with mean \code{ig_half_life_mean} and standard deviation
\code{ig_half_life_sd}. Once the duration is over, the IG "tool" is removed
from the agent, and they are again eligible for PEP if they
are exposed again.
}
\seealso{
\link[epiworldR:global-events]{epiworldR::global-events}
}
