% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cmat_symmetric.R
\name{make_cmat_symmetric}
\alias{make_cmat_symmetric}
\title{Symmetrise a contact matrix based on population sizes}
\usage{
make_cmat_symmetric(cmat, pop)
}
\arguments{
\item{cmat}{A contact matrix.}

\item{pop}{A vector of population sizes for each age group.}
}
\description{
The symmetry is achieved to ensure that \code{n(i) * c(i, j) == n(j) * c(j, i)}
where \code{n(i)} is the population size of age group \code{i} and \code{c(i, j)} is the
contact rate from age group \code{i} to age group \code{j}.
}
\details{
This function was adapted from that described in \code{socialmixr::symmetrise()}
to work with a simple matrix and population vector. The process is
implemented as described in the original function, replacing the ij-th
entries with \code{(c_ij * N_i + c_ji * N_j) / 2} and divided by the population
size of the respective age group to ensure that the resulting matrix is
symmetric with respect to the population sizes.

The resulting matrix will satisfy the condition that
\code{n(i) * c(i, j) == n(j) * c(j, i)} for all age groups \code{i} and \code{j}, ensuring
that the total number of contacts from age group \code{i} to age group \code{j} is
equal to the total number of contacts from age group \code{j} to age group \code{i}.

The operation will change the row and column sums of the contact matrix.
}
\examples{
cmat <- c(4, 2, 1, 3, 5, 6, 4, 5, 2) |>
  matrix(nrow = 3, byrow = TRUE)
pop <- c(100, 200, 300)

cmat_symm <- make_cmat_symmetric(cmat, pop)

# Checking symmetry
cmat_symm * pop == t(cmat_symm * pop)

}
